Add awards to shell games based on time played

+++ dk_interface.py
@@ -48,25 +48,37 @@ def lua_interface(emulator=None, rom=None, subfolder=None, score3=None, score2=N
         return script
 
 
-def get_award(rom, score3, score2, score1):
+def get_award(sub, rom, score3, score2, score1, time_start=None, time_end=None):
     # Read data from the compete.dat file to detemine if coins should be awarded to Jumpman.
-    try:
-        with open(COMPETE_FILE) as cf:
-            name = cf.readline().strip()
-            score = cf.readline().strip()
-        os.remove(COMPETE_FILE)
-        if rom == name and score.isnumeric():
-            if int(score) >= int(score1):
-                return AWARDS[2]  # Got 1st prize award
-            elif int(score) >= int(score2):
-                return AWARDS[1]  # Got 2nd prize award
-            elif int(score) >= int(score3):
-                return AWARDS[0]  # Got 3rd prize award
-            else:
-                return 0  # Got nothing
-    except (EOFError, FileNotFoundError, IOError):
+    if sub == "shell" and time_start and time_end:
+        # Award based on the amount of time the game was played in seconds
+        seconds = (time_end - time_start) + 10  # generously add 10 seconds
+        if seconds >= int(score1) * 60:
+            return AWARDS[2]  # Got 1st prize award
+        elif seconds >= int(score2) * 60:
+            return AWARDS[1]  # Got 2nd prize award
+        elif seconds >= int(score3) * 60:
+            return AWARDS[0]  # Got 3rd prize award
+        else:
+            return 0  # Got nothing
+    else:
+        try:
+            with open(COMPETE_FILE) as cf:
+                name = cf.readline().strip()
+                score = cf.readline().strip()
+            os.remove(COMPETE_FILE)
+            if rom == name and score.isnumeric():
+                if int(score) >= int(score1):
+                    return AWARDS[2]  # Got 1st prize award
+                elif int(score) >= int(score2):
+                    return AWARDS[1]  # Got 2nd prize award
+                elif int(score) >= int(score3):
+                    return AWARDS[0]  # Got 3rd prize award
+                else:
+                    return 0  # Got nothing
+        except (EOFError, FileNotFoundError, IOError):
+            return 0
         return 0
-    return 0
 
 
 def format_K(number, lowest_number=None):
